+++ tests/wps_restapi/test_processes.py
@@ -995,6 +995,87 @@ class WpsRestApiProcessesTest(unittest.TestCase):
                 "formats": [{"default": True, "mediaType": "text/plain"}]
             }]
 
+    def test_deploy_process_CWL_CudaRequirement_executionUnit(self):
+        with contextlib.ExitStack() as stack:
+            stack.enter_context(mocked_wps_output(self.settings))
+            cuda_requirements = {
+                "cudaVersionMin": "11.4",
+                "cudaComputeCapability": "3.0",
+                "cudaDeviceCountMin": 1,
+                "cudaDeviceCountMax": 8
+            }
+            docker_requirement = {"dockerPull": "python:3.7-alpine"}
+            cwl = {
+                "class": "CommandLineTool",
+                "cwlVersion": "v1.2",
+                "hints": {
+                    "cwltool:CUDARequirement": cuda_requirements,
+                    "DockerRequirement": docker_requirement
+                },
+                "$namespaces": {
+                    "cwltool": "http://commonwl.org/cwltool#"
+                },
+                "inputs": {},
+                "outputs": {
+                    "output": {
+                        "type": "File",
+                        "outputBinding": {
+                            "glob": "stdout.log"
+                        },
+                    }
+                }
+            }
+
+            p_id = "test-cuda"
+            body = {
+                "processDescription": {"process": {"id": p_id}},
+                "executionUnit": [{"unit": cwl}],
+                "deploymentProfileName": "http://www.opengis.net/profiles/eoc/dockerizedApplication",
+            }
+            desc = self.deploy_process_make_visible_and_fetch_deployed(body, p_id, assert_io=False)
+            pkg = self.get_application_package(p_id)
+            assert desc["deploymentProfile"] == "http://www.opengis.net/profiles/eoc/dockerizedApplication"
+            assert desc["process"]["id"] == p_id
+            assert pkg["hints"]["cwltool:CUDARequirement"] == cuda_requirements
+            assert pkg["hints"]["DockerRequirement"] == docker_requirement
+
+    def test_deploy_process_CWL_NetworkRequirement_executionUnit(self):
+        with contextlib.ExitStack() as stack:
+            stack.enter_context(mocked_wps_output(self.settings))
+            network_access_requirement = {"networkAccess": True}
+            docker_requirement = {"dockerPull": "python:3.7-alpine"}
+            for type in ["hints", "requirements"]:
+                cwl = {
+                    "class": "CommandLineTool",
+                    "cwlVersion": "v1.2",
+                    type: {
+                        "NetworkAccess": network_access_requirement,
+                        "DockerRequirement": docker_requirement
+                    },
+                    "inputs": {},
+                    "outputs": {
+                        "output": {
+                            "type": "File",
+                            "outputBinding": {
+                                "glob": "stdout.log"
+                            },
+                        }
+                    }
+                }
+
+                p_id = "test-network-access-" + type
+                body = {
+                    "processDescription": {"process": {"id": p_id}},
+                    "executionUnit": [{"unit": cwl}],
+                    "deploymentProfileName": "http://www.opengis.net/profiles/eoc/dockerizedApplication",
+                }
+                desc = self.deploy_process_make_visible_and_fetch_deployed(body, p_id, assert_io=False)
+                pkg = self.get_application_package(p_id)
+                assert desc["deploymentProfile"] == "http://www.opengis.net/profiles/eoc/dockerizedApplication"
+                assert desc["process"]["id"] == p_id
+                assert pkg[type]["NetworkAccess"] == network_access_requirement
+                assert pkg[type]["DockerRequirement"] == docker_requirement
+
     @mocked_remote_server_requests_wps1([
         resources.TEST_REMOTE_SERVER_URL,
         resources.TEST_REMOTE_SERVER_WPS1_GETCAP_XML,